requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
arxiv==2.3.1
async-timeout==5.0.1
attrs==25.4.0
av==16.1.0
beautifulsoup4==4.14.3
boto3==1.42.25
botocore==1.42.25
brotli==1.2.0
bs4==0.0.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
datasets==4.4.2
dill==0.4.0
distro==1.9.0
doclayout_yolo==0.0.4
duckduckgo_search==8.1.1
exceptiongroup==1.3.0
faiss-cpu==1.13.2
fast-langdetect==0.2.5
fastapi==0.128.0
fasttext-predict==0.9.2.4
feedparser==6.0.12
ffmpy==1.0.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
gradio==5.49.1
gradio_client==1.13.3
gradio_pdf==0.0.22
groovy==0.1.2
gvar==13.1.7
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-retries==0.4.5
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
igraph==1.0.0
ImageIO==2.37.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
json_repair==0.55.0
jsonref==1.1.0
julia==0.6.2
kiwisolver==1.4.9
lazy_loader==0.4
loguru==0.7.3
lsqfit==13.3.1
lxml==6.0.2
magika==1.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mineru==2.7.1
mineru_vl_utils==0.1.20
mlx==0.30.1
mlx-lm==0.29.1
mlx-metal==0.30.1
mlx-vlm==0.3.9
modelscope==1.33.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.4.2
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.23.2
openai==2.8.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pdfminer==20191125
pdfminer.six==20260107
pdftext==0.6.3
pillow==11.3.0
platformdirs==4.5.0
polars==1.37.0
polars-runtime-32==1.37.0
primp==0.15.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.2.1
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.26.6
pyparsing==3.2.5
pypdf==6.6.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
RapidFuzz==3.14.3
regex==2025.11.3
reportlab==4.4.7
requests==2.32.5
rich==14.2.0
robust-downloader==0.0.2
ruff==0.14.11
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.2.0
sentencepiece==0.2.1
sentry-sdk==2.46.0
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
starlette==0.50.0
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.2
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.252
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.40.0
vegas==6.3
wandb==0.23.0
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
